First pass at bringing in LocatedA for API anns in locations
Treatment of ECP in parsing is provisional at this stage, leads to some
horribly stuff in Parser.y and RdrHsSyn.

It is an extensive but not invasive change. I think (AZ).

It only introduces the type and works it through, does not actually
populate it with any annotations.

Locally it reports some parsing tests using less memory.